[SUS-46] Refactor/init_refactor
