+++ src/core/error/ErrorRegistry.ts
