+++ src/core/util/nickname/const/nicknameList.ts
