--- src/dto/frontend/request/EmailSendReques.ts
@@ -1,1 +1,0 @@
-interface EmailSendRequestParams {}
