--- src/dto/frontend/request/SendFindPasswordEmailRequest.ts
@@ -1,14 +1,0 @@
-interface SendFindPasswordEmailRequestParams {
-  id: string
-  email: string
-}
-
-export default class SendFindPasswordEmailRequest {
-  public id: string
-  public email: string
-
-  constructor(parms: SendFindPasswordEmailRequestParams) {
-    this.id = parms.id
-    this.email = parms.email
-  }
-}
