--- src/dto/frontend/request/SendVerifyEmailRequest.ts
@@ -1,14 +1,0 @@
-interface SendVerifyEmailRequestParams {
-  id: string
-  email: string
-}
-
-export default class SendVerifyEmailRequest {
-  public id: string
-  public email: string
-
-  constructor(parms: SendVerifyEmailRequestParams) {
-    this.id = parms.id
-    this.email = parms.email
-  }
-}
