+++ src/member/dao/frontend/request/SendFindPasswordEmailRequest.ts
@@ -1,0 +1,14 @@
+interface SendFindPasswordEmailRequestParams {
+  id: string
+  email: string
+}
+
+export default class SendFindPasswordEmailRequest {
+  public id: string
+  public email: string
+
+  constructor(parms: SendFindPasswordEmailRequestParams) {
+    this.id = parms.id
+    this.email = parms.email
+  }
+}
